pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "thoughtstream"
version = "0.1.0"
description = "Intelligent context-aware note-taking daemon for developers."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "ThoughtStream Authors" }
]
keywords = ["notes", "productivity", "llm", "developer-tools"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Intended Audience :: Developers",
  "Topic :: Office/Business :: Scheduling"
]
dependencies = [
  "rich>=13.0.0",
  "typer>=0.12.0",
  "pynput>=1.7.0",
  "pyperclip>=1.8.0",
  "ollama>=0.1.0",
  "sqlite-vss>=0.1.2",
  "pydantic>=2.0.0",
  "orjson>=3.9.0",
  "PyYAML>=6.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "ruff>=0.6.0",
  "mypy>=1.8.0"
]

[project.urls]
"Homepage" = "https://github.com/kwizzlesurp10-ctrl/thoughtstream"
"Documentation" = "https://github.com/kwizzlesurp10-ctrl/thoughtstream/docs"
"Source" = "https://github.com/kwizzlesurp10-ctrl/thoughtstream"

[project.scripts]
thoughtstream = "thoughtstream.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

